REF: solver_from_rules_set is now a proper MiniSATSolver ctor.
